Removed CSV namespacing from Parser and SyncParser types. Removed CSV.Delimiter enum
