+++ Sources/CSV/CSV.swift
@@ -10,11 +10,4 @@ public struct CSV {
             self.fields = fields
         }
     }
-    
-    internal struct Delimiter {
-        static let comma = UInt8(ascii: ",")
-        static let quote = UInt8(ascii: "\"")
-        static let newLine = UInt8(ascii: "\n")
-        static let carriageReturn = UInt8(ascii: "\r")
-    }
 }
